tried optimizing cxrloader

+++ CheXpert2/dataloaders/CXRLoader.py
@@ -10,7 +10,6 @@ import copy
 
 import cv2 as cv
 import numpy as np
-import pandas as pd
 import torch
 import tqdm
 import yaml
@@ -87,8 +86,9 @@ class CXRLoader(Dataset):
 
 
         self.dataset = dataset
-        self.files = MongoDB("10.128.107.212", 27017, ["ChexPert", "ChexNet", "ChexXRay"]).dataset(dataset,classnames=classnames)
-
+        self.files = MongoDB("10.128.107.212", 27017, ["ChexPert", "ChexNet", "ChexXRay"]).dataset(dataset,
+                                                                                                   classnames=classnames)
+        self.files[self.classes] = self.files[self.classes].astype(int)
         self.img_dir = img_dir
